requirements.txt
fastapi
python-multipart
jinja2
openai
python-dotenv
uvicorn
gunicorn
fastapi-mail